part1-database-etl/requirements.txt
pandas
numpy
mysql-connector-python
python-dateutil
